started reworking system files to cpp with config struct.
